+++ packages/libp2p-daemon/CHANGELOG.md
@@ -4,6 +4,80 @@
 
 * update all deps ([#348](https://github.com/libp2p/js-libp2p-daemon/issues/348)) ([44bf148](https://github.com/libp2p/js-libp2p-daemon/commit/44bf148f37b9f4091dc8ed26cf343e196607ebbc))
 
+## [6.0.30](https://github.com/libp2p/js-libp2p/compare/daemon-v6.0.29...daemon-v6.0.30) (2026-06-13)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/daemon-server bumped from ^9.0.29 to ^9.0.30
+
+## [6.0.29](https://github.com/libp2p/js-libp2p/compare/daemon-v6.0.28...daemon-v6.0.29) (2026-06-06)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/daemon-server bumped from ^9.0.28 to ^9.0.29
+
+## [6.0.28](https://github.com/libp2p/js-libp2p/compare/daemon-v6.0.27...daemon-v6.0.28) (2026-05-30)
+
+
+### Dependencies
+
+* upgrade multiformats to 14 and multiaddr to 13.0.3 ([#3526](https://github.com/libp2p/js-libp2p/issues/3526)) ([5b8813a](https://github.com/libp2p/js-libp2p/commit/5b8813abcbca3fd59e6d137f2a42ff53e115cf99))
+* upgrade to aegir v48 ([#3495](https://github.com/libp2p/js-libp2p/issues/3495)) ([29797a5](https://github.com/libp2p/js-libp2p/commit/29797a5bbbfc5b93d91fcdc244d7f65e2ce5e76c))
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/daemon-server bumped from ^9.0.27 to ^9.0.28
+
+## [6.0.27](https://github.com/libp2p/js-libp2p/compare/daemon-v6.0.26...daemon-v6.0.27) (2026-05-27)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/daemon-server bumped from ^9.0.26 to ^9.0.27
+
+## [6.0.26](https://github.com/libp2p/js-libp2p/compare/daemon-v6.0.25...daemon-v6.0.26) (2026-05-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/daemon-server bumped from ^9.0.25 to ^9.0.26
+
+## [6.0.25](https://github.com/libp2p/js-libp2p/compare/daemon-v6.0.24...daemon-v6.0.25) (2026-05-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/daemon-server bumped from ^9.0.24 to ^9.0.25
+
+## [6.0.24](https://github.com/libp2p/js-libp2p/compare/daemon-v6.0.23...daemon-v6.0.24) (2026-05-13)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/daemon-server bumped from ^9.0.23 to ^9.0.24
+
+## [6.0.23](https://github.com/libp2p/js-libp2p/compare/daemon-v6.0.22...daemon-v6.0.23) (2026-05-09)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/daemon-server bumped from ^9.0.22 to ^9.0.23
+
 ## [6.0.22](https://github.com/libp2p/js-libp2p/compare/daemon-v6.0.21...daemon-v6.0.22) (2026-05-04)